+++ src/components/ProviderInitializer/ProviderInitializer.tsx
@@ -1,7 +1,4 @@
-import { useEffect, useRef, useState } from 'react';
-import { ExtensionProvider } from '@multiversx/sdk-extension-provider';
-import { HWProvider } from '@multiversx/sdk-hw-provider';
-import { OperaProvider } from '@multiversx/sdk-opera-provider';
+import { useEffect, useRef } from 'react';
 import { getNetworkConfigFromApi } from 'apiCalls';
 import { useLoginService } from 'hooks/login/useLoginService';
 import { useWalletConnectV2Login } from 'hooks/login/useWalletConnectV2Login';
@@ -21,8 +18,7 @@ import {
   walletLoginSelector,
   ledgerLoginSelector,
   isLoggedInSelector,
-  tokenLoginSelector,
-  logoutRouteSelector
+  tokenLoginSelector
 } from 'reduxStore/selectors/loginInfoSelectors';
 import { networkSelector } from 'reduxStore/selectors/networkConfigSelectors';
 import {
@@ -40,12 +36,12 @@ import {
   getAccount,
   getLatestNonce,
   newWalletProvider,
-  getLedgerConfiguration,
   emptyProvider
 } from 'utils/account';
-import { logout } from 'utils/logout';
 import { parseNavigationParams } from 'utils/parseNavigationParams';
-import { useWebViewLogin } from '../hooks/login/useWebViewLogin';
+import { useWebViewLogin } from '../../hooks/login/useWebViewLogin';
+import { getExtensionProvider } from './helpers';
+import { useSetLedgerProvider } from './hooks';
 
 let initalizingLedger = false;
 
@@ -58,12 +54,7 @@ export function ProviderInitializer() {
   const ledgerAccount = useSelector(ledgerAccountSelector);
   const ledgerLogin = useSelector(ledgerLoginSelector);
   const isLoggedIn = useSelector(isLoggedInSelector);
-  const logoutRoute = useSelector(logoutRouteSelector);
-  const [ledgerData, setLedgerData] =
-    useState<{
-      version: string;
-      dataEnabled: boolean;
-    }>();
+
   const tokenLogin = useSelector(tokenLoginSelector);
   const nativeAuthConfig = tokenLogin?.nativeAuthConfig;
   const loginService = useLoginService(
@@ -71,6 +62,7 @@ export function ProviderInitializer() {
   );
   const initializedAccountRef = useRef(false);
   const dispatch = useDispatch();
+  const { setLedgerProvider, ledgerData } = useSetLedgerProvider();
 
   useWebViewLogin();
 
@@ -206,75 +198,19 @@ export function ProviderInitializer() {
     dispatch(setWalletLogin(null));
   }
 
-  async function getInitializedHwWalletProvider() {
-    const hwWalletP = new HWProvider();
-    let isInitialized = hwWalletP.isInitialized();
-    if (!isInitialized) {
-      isInitialized = await hwWalletP.init();
-    }
-    if (!isInitialized && isLoggedIn) {
-      console.warn('Could not initialise ledger app');
-      logout(logoutRoute);
-      return;
-    }
-
-    if (ledgerLogin?.index != null) {
-      await hwWalletP.setAddressIndex(ledgerLogin.index);
-    }
-
-    return hwWalletP;
-  }
-
-  async function setLedgerProvider() {
-    try {
-      const hwWalletP = await getInitializedHwWalletProvider();
-      if (!hwWalletP) {
-        return;
-      }
-      const ledgerConfig = await getLedgerConfiguration(hwWalletP);
-      setAccountProvider(hwWalletP);
-      setLedgerData(ledgerConfig);
-    } catch (err) {
-      console.error('Could not initialise ledger app', err);
-      logout(logoutRoute);
-    }
-  }
-
   async function setExtensionProvider() {
-    try {
-      const address = await getAddress();
-      const provider = ExtensionProvider.getInstance().setAddress(address);
-
-      const success = await provider.init();
-
-      if (success) {
-        setAccountProvider(provider);
-      } else {
-        console.error(
-          'Could not initialise extension, make sure MultiversX wallet extension is installed.'
-        );
-      }
-    } catch (err) {
-      console.error('Unable to login to ExtensionProvider', err);
+    const address = await getAddress();
+    const provider = await getExtensionProvider(address);
+    if (provider) {
+      setAccountProvider(provider);
     }
   }
 
   async function setOperaProvider() {
-    try {
-      const address = await getAddress();
-      const provider = OperaProvider.getInstance().setAddress(address);
-
-      const success = await provider.init();
-
-      if (success) {
-        setAccountProvider(provider);
-      } else {
-        console.error(
-          'Could not initialise opera crypto wallet, make sure that opera crypto wallet is installed.'
-        );
-      }
-    } catch (err) {
-      console.error('Unable to login to OperaProvider', err);
+    const address = await getAddress();
+    const provider = await getExtensionProvider(address);
+    if (provider) {
+      setAccountProvider(provider);
     }
   }