+++ src/components/ProviderInitializer/helpers/index.ts
@@ -1,0 +1,2 @@
+export * from './getOperaProvider';
+export * from './getExtensionProvider';