+++ src/hooks/transactions/helpers/index.ts
@@ -2,3 +2,4 @@ export * from './getShouldMoveTransactionsToSignedState';
 export * from './useSetTransactionNonces';
 export * from './getAreAllTransactionsSignedByGuardian';
 export * from './checkNeedsGuardianSigning';
+export * from './useGetLedgerProvider';