+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [Fixed ledger signing reconnect](https://github.com/multiversx/mx-sdk-dapp/pull/874)
 - [Added ledger transport mock](https://github.com/multiversx/mx-sdk-dapp/pull/871)
 - [Fixed clear navigation url after signing a message](https://github.com/multiversx/mx-sdk-dapp/pull/872)